Expose TicksGenerator mocks for testing
Put TicksGenerator mocks in a separate header file for use in other
tests.

Introduce TicksGenerator::isGenerating() for testing purposes.
